requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==21.2.0
autopep8==1.6.0
boto3==1.20.2
botocore==1.23.2
cattrs==1.8.0
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==2.0.7
cloudinary==1.28.0
coverage==5.2.1
dj-database-url==0.5.0
Django==3.0.7
django-common-helpers==0.9.2
django-cors-headers==3.4.0
django-cron==0.5.1
django-crontab==0.7.1
django-debug-toolbar==2.2
django-filter==2.3.0
django-storages==1.12.3
djangorestframework==3.11.0
djangorestframework-simplejwt==5.0.0
gunicorn==20.0.4
hellosign-python-sdk==4.0.0
idna==2.10
importlib-metadata==1.7.0
isort==5.6.4
jmespath==0.10.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
mysqlclient==2.1.0
protobuf==3.18.1
psycopg2-binary==2.8.6
pycodestyle==2.8.0
pycryptodome==3.11.0
PyJWT==2.2.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-decouple==3.3
python-http-client==3.3.3
pytz==2020.1
rave-python==1.2.13
requests==2.24.0
requests-cache==0.5.2
s3transfer==0.5.0
sendgrid==6.6.0
six==1.15.0
sqlparse==0.3.1
starkbank-ecdsa==2.0.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==4.0.1
url-normalize==1.4.3
urllib3==1.25.10
whitenoise==5.1.0
wrapt==1.12.1
zipp==3.1.0
